Move all R stuff to opad, since its only required there
